requirements.txt
altair==6.0.0
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
Automat==22.10.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
boto3==1.34.46
botocore==1.34.46
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.6
click==8.3.1
cloud-init==24.1.3
colorama==0.4.6
comm==0.2.3
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
docstring_parser==0.17.0
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==22.10.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.0.1
json5==0.13.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lark==1.3.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.4.3
oauthlib==3.2.2
openai==2.29.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.2.2
python-json-logger==4.0.0
python-magic==0.4.27
pytz==2026.1.post1
PyYAML==6.0.1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.7.1
rpds-py==0.30.0
s3transfer==0.10.1
Send2Trash==2.1.0
service-identity==24.1.0
setuptools==68.1.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sos==4.10.2
soupsieve==2.8.3
ssh-import-id==5.11
stack-data==0.6.3
streamlit==1.55.0
systemd-python==235
tenacity==9.1.4
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
Twisted==24.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.6.3
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.42.0
widgetsnbextension==4.0.15
zope.interface==6.1
